Update EC_Processing.R with Parliaments 36 and 37
This script is still a work in progress. This commit saves a recent addition to the script that processes election results from parliaments 36 and 37. The raw data from Elections Canada are presented in messy tab-delimited text files. The current script wrangles the data into a tidier format. Additional data processing is required before data on these years are suitable for analysis.